convert all money from decimal to int

+++ CinemaTicket/Entities/OnlineEntities.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public UserEntities UserId { get; set; }
         public ShoppingBagEntities ShoppingBagId { get; set; }
-        public decimal TransAmount { get; set; }
+        public int TransAmount { get; set; }
         public int OrderNo { get; set; }
         public int TransNo { get; set; }
         public DateTime CreateDate { get; set; }
